fix : fix the folder structure in the front end refs : MBM-05-refactor-react-components
