+++ reactfront/src/components/Home/ShowVecinos.css
@@ -27,7 +27,7 @@ body {
   }
 .hero {
     height: 550px;
-    background-image: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)), url("./../img/building.jpg");
+    background-image: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)), url("./../../img/building.jpg");
     background-position: center;
     background-repeat: no-repeat;
     background-size: cover;
